Block editor: fix focus handler in Safari
